+++ src/app/components/split-view/multi/data/terra-multi-split-view.config.ts
@@ -4,6 +4,7 @@ import {
 } from '@angular/core';
 import { TerraMultiSplitViewInterface } from './terra-multi-split-view.interface';
 import { isNullOrUndefined } from 'util';
+import * as CircularJSON from 'circular-json';
 
 @Injectable()
 export class TerraMultiSplitViewConfig
@@ -62,12 +63,14 @@ export class TerraMultiSplitViewConfig
                         for(let child of parent.children)
                         {
                             // TODO very ugly way, maybe add an option to use an id?
-                            let hasSameParameter:boolean =
-                                (child.parameter && view.parameter && JSON.stringify(child.parameter) === JSON.stringify(view.parameter)) ||
-                                (child.inputs && view.inputs && JSON.stringify(child.inputs) === JSON.stringify(view.inputs)) ||
-                                (child.name === view.name);
+                            let hasSameParameter:boolean = child.parameter && view.parameter &&
+                                                           CircularJSON.stringify(child.parameter) === CircularJSON.stringify(view.parameter);
+                            let hasSameInputs:boolean = child.inputs && view.inputs &&
+                                                        CircularJSON.stringify(child.inputs) === CircularJSON.stringify(view.inputs);
+                            let hasSameName:boolean = child.name === view.name;
+                            let hasSameModule:boolean = child.module.ngModule === view.module.ngModule;
 
-                            if(hasSameParameter && child.module.ngModule === view.module.ngModule)
+                            if(hasSameModule && (hasSameParameter || hasSameInputs || hasSameName))
                             {
                                 view = child;
                                 viewExist = true;